Create a base fragment class
